+++ gradle/libs.versions.toml
@@ -1,6 +1,6 @@
 [versions]
 junit = "5.10.0"
-slf4j = "2.0.7"
+slf4j = "2.0.9"
 log4j2 = "2.20.0"
 
 [libraries]
