+++ paper.md
@@ -30,7 +30,7 @@ The general-purpose network analysis software `UCINET` [@bef-uw-02], `visone` [@
 
 Nonetheless, only few tools offer partial support to compute a small selection of non-relaxed role equivalences, yet these solutions do not generalize to other kinds of role equivalences and are restricted to specific kinds of networks, such as unweighted networks or networks with categorical tie weights.
 
-The `netroles` library is intended to close this gap and aims to extend the aim of applicability of role equivalences. It provides a unified framework for role equivalences based on the formalization of role equivalence proposed in @mb-er-22. The library implements established notions of role, but its main goal is to enable the expression of novel and nuanced concepts of role, incorporating vertex and edge attributes as well as multiple relations according to the requirements of the network study. This also facilities the specification of error-tolerant role equivalences that allow for minor deviations from the ideal case. The example analyses in @mb-er-22 were conducted using `netroles`.
+The `netroles` library is intended to close this gap and aims to extend the aim of applicability of role equivalences. It provides a unified framework for role equivalences based on the formalization of role equivalence proposed in @mb-er-22. The library implements established notions of role, but its main goal is to enable the expression of new role concepts, incorporating vertex and edge attributes as well as multiple relations according to the requirements of the network study. This also facilities the specification of error-tolerant role equivalences that allow for minor deviations from the ideal case. The example analyses in @mb-er-22 were conducted using `netroles`.
 
 # Background
 
