### 2023-09-10 #5812 @muellerj2
I'm sorry, I didn't know that the editorialbot expects to find the software on the paper branch as well.
I rebased the paper branch on top of the main branch now so that all statistics are obtained immediately in the review issue.
