### 2023-09-15 #5812 @muellerj2
It's difficult to find someone knowledge on this research direction and versed in Java nowadays, but let me suggest the following two reviewers for now:
* Philip Leifeld (github handle leifeld)
* Sébastien Heymann (github handle sheymann), co-founder of Gephi
