### 2023-09-17 #5812 @fboehm
@leifeld @sheymann - would you be willing to review this submission for JOSS? At JOSS, our reviews are checklist-driven. For example, here is a checklist for another JOSS submission: https://github.com/openjournals/joss-reviews/issues/5754#issuecomment-1674317863
Thank you for considering this invitation. Please let me know if you have any questions.
