### 2023-09-25 #5812 @fboehm
@sdmccabe - would you be willing to review this submission for JOSS? Thank you!
