### 2023-09-27 #5812 @fboehm
@tobias-dv-lnu - would you be willing to review this submission for JOSS?

Thank you !
