### 2023-09-29 #5812 @abhishektiwari
@fboehm I am open to review this submission. I have now reviewed a few network analysis packages for JOSS. Please feel free to assign me as reviewer.
