### 2023-09-30 #5903 @editorialbot
[REVIEW]: netroles: A Java library for role equivalence analysis in networks
**Submitting author:** @muellerj2 (<a href="http://orcid.org/0000-0001-6440-8046">Julian Müller</a>)
**Repository:** https://github.com/muellerj2/netroles
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0
