### 2023-10-03 c3a7c53
improve scalability of progress reporting for parallel computations
