### 2023-10-17 4bbfe84
Bump log4j2 from 2.20.0 to 2.21.0
Bumps `log4j2` from 2.20.0 to 2.21.0.

Updates `org.apache.logging.log4j:log4j-slf4j2-impl` from 2.20.0 to 2.21.0

Updates `org.apache.logging.log4j:log4j-core` from 2.20.0 to 2.21.0

---
updated-dependencies:
- dependency-name: org.apache.logging.log4j:log4j-slf4j2-impl
  dependency-type: direct:production
  update-type: version-update:semver-minor
- dependency-name: org.apache.logging.log4j:log4j-core
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
