+++ gradle/libs.versions.toml
@@ -1,7 +1,7 @@
 [versions]
 junit = "5.10.0"
 slf4j = "2.0.9"
-log4j2 = "2.20.0"
+log4j2 = "2.21.0"
 
 [libraries]
 junit-launcher = { group = "org.junit.platform", name = "junit-platform-launcher" }
