+++ gradle/libs.versions.toml
@@ -10,7 +10,7 @@ slf4j-api = { group = "org.slf4j", name = "slf4j-api", version.ref = "slf4j" }
 log4j2-slf4j-bridge = { group = "org.apache.logging.log4j", name = "log4j-slf4j2-impl", version.ref = "log4j2" }
 log4j2-core = { group = "org.apache.logging.log4j", name = "log4j-core", version.ref = "log4j2" }
 json = { group = "org.json", name = "json", version = "20230618" }
-jackson-databind = { group = "com.fasterxml.jackson.core", name = "jackson-databind", version = "2.15.2" }
+jackson-databind = { group = "com.fasterxml.jackson.core", name = "jackson-databind", version = "2.15.3" }
 univocity-parsers = { group = "com.univocity", name = "univocity-parsers", version = "2.9.1" }
 
 [bundles]
