### 2023-10-21 muellerj2/netroles#33 @abhishektiwari
Link Javadoc for API Documentation
Although software is documented to a satisfactory level, for a Java package API documentation is scattered between `DESIGN.md` and `USAGE.md` and not very structured. Using an API documentation generator like Javadoc or something similar is strongly recommended. It seems you already using Javadoc comments, so it should be easier to generate. Or maybe you are generating Javadoc but not published it yet.

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/5903)

Update:
See comments below.
