### 2023-10-21 #5903 @abhishektiwari
@muellerj2 Please see my initial feedback. I am still working to verify the functional claims of the software.

## General checks

### Substantial scholarly effort
`netroles` is relatively new software with GitHub project age just over 3 months, 40 GitHub commits, 101991 LOC, and GitHub commit history starting from Jul 23, 2023.

Project meets borderline well-established software project criteria established by JOSS. Need additional confirmation from author.

— Age of software on GitHub is just over 3 months, although it seems software was imported from somewhere else ([first commit](https://github.com/muellerj2/netroles/commit/a87e966149a842d28b03d86d6b07b292ab77f33e))
— Software is used to perform an analysis for a peer revived paper published by *Müller & Brandes* (2022)
— Total lines of code (LOC) 101991 including, Java LOC 93566.
— 40+ commits on GitHub
— Based on software review, software is useful and is likely used by others for their research

@mullerj2 can you confirm how long software has been under active development before publishing to GitHub?

## Documentation

### Functionality documentation
Although software is documented to a satisfactory level, for a Java package API documentation is scattered between `DESIGN.md` and `USAGE.md` and not very structured. Using an API documentation generator like Javadoc or something similar is strongly recommended. It seems author is already using Javadoc comments, so it should be easier to generate. Or maybe they are generating Javadoc but not published it yet ([1](https://github.com/muellerj2/netroles/issues/33)).

## Software paper

### Quality of writing
`formalizations of the concepts` or `formalization of the concepts`. Grammatically later.
