### 2023-10-21 muellerj2/netroles#33 @abhishektiwari
Given project released to [Maven Central](https://search.maven.org/) then your Javadoc is available at [javadoc.io](http://www.javadoc.io/) automatically. Can you link your Documentation/README page the javadoc.io?

https://www.javadoc.io/doc/io.github.muellerj2/netroles-api/latest/index.html
https://www.javadoc.io/doc/io.github.muellerj2/netroles-engine/latest/index.html
https://www.javadoc.io/doc/io.github.muellerj2/netroles-io-api/latest/index.html
https://www.javadoc.io/doc/io.github.muellerj2/netroles-io-engine/latest/index.html
https://www.javadoc.io/doc/io.github.muellerj2/netroles-lang-api/latest/index.html
https://www.javadoc.io/doc/io.github.muellerj2/netroles-networks-engine/latest/index.html
https://www.javadoc.io/doc/io.github.muellerj2/netroles-networks-api/latest/index.html

You can also generate Javadoc badges like

[![javadoc](https://javadoc.io/badge2/io.github.muellerj2/netroles-networks-engine/javadoc.svg)](https://javadoc.io/doc/io.github.muellerj2/netroles-networks-engine)
