### 2023-10-23 muellerj2/netroles#33 @leifeld
I second this. I would either link to the Maven docs or your own generated javadoc in the Markdown files here in the repo.
