### 2023-10-23 muellerj2/netroles#34 @leifeld
SVG with transparent background in linked document
The [SVG diagram](https://github.com/muellerj2/netroles/raw/main/docs/bidirectional-roleoperator-dataflow.svg) in the [linked document](https://github.com/muellerj2/netroles/blob/main/docs/DESIGN.md) has dark foreground on a transparent background. This makes it hard to read in a dark theme. I would add a non-transparent background.

As an aside, I would not call it "linked document" but rather by its actual title when you refer to it in the README.

[JOSS review](https://github.com/openjournals/joss-reviews/issues/5903)
