### 2023-10-24 a2803c3
Add backtrack fixed point enumeration algorithm
