+++ modules/roles-api/src/main/java/ch/ethz/sn/visone3/roles/lattice/BacktrackSearchEnumerator.java
@@ -1,0 +1,279 @@
+/*
+ * This file is part of netroles.
+ *
+ * netroles is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * netroles is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with netroles.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package ch.ethz.sn.visone3.roles.lattice;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.function.Function;
+import java.util.function.Supplier;
+
+/**
+ * This enumeration algorithm searches fixed points of a monotone function on a
+ * specific underlying lattice by exploiting that the elements of a lattice can
+ * be decomposed into several dimensions, so that projections of a lattice
+ * element to a given set of dimensions can be obtained as well as extensions of
+ * a given projection to more dimensions, or the set of extremal extensions of a
+ * projection to full lattice elements under the lattice ordering can be listed.
+ * 
+ * <p>
+ * This enumeration algorithm thus needs at least the following ingredients
+ * besides the monotone function f:
+ * <ul>
+ * <li>a way to decompose lattice elements into several dimensions, including an
+ * easy way to project to a lower number of dimensions,</li>
+ * <li>an iterable that produces all extensions of a projection by another
+ * dimension, as well as</li>
+ * <li>an iterable over all extremal extensions of a projection to full lattice
+ * elements under the lattice ordering.</li>
+ * </ul>
+ * 
+ * Fixed points are then essentially found using the following algorithm for
+ * monotone function f:
+ * 
+ * <pre>
+ * function search(projection, nextdim, maxdim)
+ *   if nextdim > maxdim then
+ *     if f(projection) == projection then
+ *       output projection;
+ *     end if
+ *   else
+ *     for all extensions e of projection by dimension nextdim do
+ *       if there is an extremal extension m of e to a full lattice element
+ *           such that e is a projection of f(m) then
+ *         search(e, nextdim + 1, maxdim);
+ *       end if
+ *     end for
+ *   end if
+ * end function
+ * 
+ * // initial call to start enumeration 
+ * search(0-dimension projection, 1, n)
+ * </pre>
+ * 
+ * Generally, the algorithm has to do little bookkeeping, and it performs well
+ * under two conditions:
+ * <ul>
+ * <li>if it is easy to generate another extension of a projection by a single
+ * dimension, and</li>
+ * <li>if there are only few extremal extensions of a projection to a full
+ * lattice element, and these are easy to generate.</li>
+ * </ul>
+ * 
+ * These conditions tend to be satisfied if the meet (if f is increasing
+ * monotone) or join (if f is decreasing monotone) act on the decomposition of
+ * lattice elements into dimensions essentially dimension-wise. For example,
+ * this is the case on the lattice of binary relations, where meet and join
+ * correspond to set intersection/union (if binary relations are modeled as sets
+ * of pairs) or component-wise minimum/maximum (if binary relations are
+ * represented as binary matrices).
+ */
+public class BacktrackSearchEnumerator {
+
+  private BacktrackSearchEnumerator() {
+
+  }
+
+  /**
+   * Represents a function that accepts an object and an integer and produces a
+   * result.
[…]
