### 2023-10-24 d196f20
avoid calling JDK >= 9 method
