### 2023-10-24 db359c7
minor Javadoc change
