### 2023-10-24 cbf53a7
Add backtrack fixed point enumeration algorithm (#36)
Adds an alternative enumeration algorithm. This one tends to be faster (due to less bookkeeping overhead) when the lattice elements naturally decompose into several dimensions and this decomposition is consistent with the corresponding join and meet operations (e.g., because the join or meet operation essentially acts on each dimension independently). Specifically, the meet operation should have this structure if the fixed points of an increasing monotone function are sought, and the join operation should be like that for fixed points of a 
[…]
