+++ modules/roles-api/src/main/java/ch/ethz/sn/visone3/roles/lattice/ProjectionEnumerators.java
@@ -1,0 +1,344 @@
+/*
+ * This file is part of netroles.
+ *
+ * netroles is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * netroles is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with netroles.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package ch.ethz.sn.visone3.roles.lattice;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Iterator;
+import java.util.NoSuchElementException;
+
+import ch.ethz.sn.visone3.lang.ConstMapping;
+import ch.ethz.sn.visone3.lang.ConstMapping.OfInt;
+import ch.ethz.sn.visone3.lang.Mappings;
+import ch.ethz.sn.visone3.roles.structures.BinaryRelation;
+import ch.ethz.sn.visone3.roles.structures.BinaryRelations;
+
+/**
+ * Enumerators and other methods related to the projections of lattice elements
+ * used by the back-track search enumeration algorithm.
+ * 
+ * @see BacktrackSearchEnumerator
+ */
+public class ProjectionEnumerators {
+
+  private ProjectionEnumerators() {
+  }
+
+  /**
+   * Generates all extensions of a projection by a single dimension for binary
+   * relations.
+   * 
+   * @param projection    the projection.
+   * @param nextDimension the ordinal of the next dimension.
+   * @return an iterable over all projections when extending by another dimension.
+   */
+  public static Iterable<boolean[]> generateExtensionsBinaryRelations(boolean[] projection, int nextDimension) {
+    return () -> new Iterator<boolean[]>() {
+      private byte count = 0;
+
+      @Override
+      public boolean hasNext() {
+        return count < 2;
+      }
+
+      @Override
+      public boolean[] next() {
+        if (count >= 2) {
+          throw new NoSuchElementException();
+        }
+        projection[nextDimension] = count == 0;
+        ++count;
+        return projection;
+      }
+    };
+  }
+
+  /**
+   * Produces the extremal full binary relation extension of a projection
+   * according to the refinement ordering.
+   * 
+   * @param projection           the projection.
+   * @param numDimProjection     the number of dimensions in the projection.
+   * @param numDimBinaryRelation the number of dimensions of the binary relation
+   *                             (which is the domain size squared).
+   * @param maximal              true if the maximal extension is to be generated,
+   *                             else the minimal one is produced.
+   * @return an iterable producing the only maximal or minimal extension.
+   */
+  public static Iterable<BinaryRelation> extremalExtensionBinaryRelations(boolean[] projection, int numDimProjection,
+      int numDimBinaryRelation, boolean maximal) {
+    return () -> new Iterator<BinaryRelation>() {
+      private boolean notGenerated = true;
+
+      @Override
+      public boolean hasNext() {
+        return notGenerated;
+      }
+
+      @Override
+      public BinaryRelation next() {
+        if (!notGenerated) {
+          throw new NoSuchElementException();
+        }
+        notGenerated = false;
+        for (int i = numDimProjection; i < numDimBinaryRelation; ++i) {
+          projection[i] = maximal;
+        }
+        return projectionToBinaryRelation(projection, numDimBinaryRelation);
+      }
+    };
+  }
+
+  /**
[…]
