+++ modules/roles-impl/src/test/java/ch/ethz/sn/visone3/roles/test/lattice/LatticeTest.java
@@ -34,6 +34,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
 
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
@@ -51,7 +52,8 @@ import ch.ethz.sn.visone3.roles.blocks.RoleOperator;
 import ch.ethz.sn.visone3.roles.blocks.RoleOperators;
 import ch.ethz.sn.visone3.roles.impl.algorithms.Equivalences;
 import ch.ethz.sn.visone3.roles.lattice.CoverEnumerators;
-import ch.ethz.sn.visone3.roles.lattice.FixedPointEnumerator.CoverEnumerator;
+import ch.ethz.sn.visone3.roles.lattice.DepthFirstSearchEnumerator.CoverEnumerator;
+import ch.ethz.sn.visone3.roles.lattice.ProjectionEnumerators;
 import ch.ethz.sn.visone3.roles.lattice.StableRolesEnumeration;
 import ch.ethz.sn.visone3.roles.position.NetworkView;
 import ch.ethz.sn.visone3.roles.structures.BinaryRelation;
@@ -62,7 +64,8 @@ import ch.ethz.sn.visone3.roles.structures.RelationBuilders;
 
 public class LatticeTest {
 
-  // Political actors network from Doreian and Albert - Partitioning Political Actor Networks: Some
+  // Political actors network from Doreian and Albert - Partitioning Political
+  // Actor Networks: Some
   // Quantitative Tools for Analyzing Qualitative Networks
 
   private static final Integer z = null;
@@ -83,16 +86,23 @@ public class LatticeTest {
       { z, z, 1, z, z, 1, z, z, z, z, z, z, z, z }, //
   };
 
+  private static final Integer[][] cycleNetwork = { //
+      { z }, //
+      { 1, z }, //
+      { z, 1, z }, //
+      { z, z, 1, z }, //
+      { z, z, z, 1, z }, //
+      { z, z, z, z, 1, z }, //
+      { 1, z, z, z, z, 1, z }, //
+  };
+
   @Test
   public void testRegularEquivalencePoliticalActors() throws IOException {
     Network net = MatrixSource.fromAdjacency(politicalActorsNetwork, false).getNetwork();
 
-    Iterable<ConstMapping.OfInt> fixedPointsIterable = StableRolesEnumeration.EQUIVALENCE
-        .stableRolesUnderRestriction(
-            RoleOperators.EQUIVALENCE.regular()
-                .of(NetworkView.fromNetworkRelation(net, Direction.OUTGOING))
-                .make(),
-            Converters.singleClassEquivalence(net.countMonadicIndices()).apply(null));
+    Iterable<ConstMapping.OfInt> fixedPointsIterable = StableRolesEnumeration.EQUIVALENCE.stableRolesUnderRestriction(
+        RoleOperators.EQUIVALENCE.regular().of(NetworkView.fromNetworkRelation(net, Direction.OUTGOING)).make(),
+        Converters.singleClassEquivalence(net.countMonadicIndices()).apply(null));
 
     int count = 0;
     Iterator<ConstMapping.OfInt> iterator = fixedPointsIterable.iterator();
@@ -110,23 +120,15 @@ public class LatticeTest {
   public void testErrorTolerantExactEquivalencePoliticalActors() throws IOException {
     Network net = MatrixSource.fromAdjacency(politicalActorsNetwork, false).getNetwork();
 
-    RoleOperator<ConstMapping.OfInt> roleOp = Operators
-        .composeRoleOp(
-            Operators
-                .composeConv(
-                    Operators
-                        .composeConv(
[…]
