### 2023-10-24 e9d30ff
Bump log4j2 from 2.21.0 to 2.21.1 (#35)
Bumps `log4j2` from 2.21.0 to 2.21.1.

Updates `org.apache.logging.log4j:log4j-slf4j2-impl` from 2.21.0 to 2.21.1

Updates `org.apache.logging.log4j:log4j-core` from 2.21.0 to 2.21.1

---
updated-dependencies:
- dependency-name: org.apache.logging.log4j:log4j-slf4j2-impl
  dependency-type: direct:production
  update-type: version-update:semver-patch
- dependency-name: org.apache.logging.log4j:log4j-core
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
