### 2023-10-25 d8fc2d4
change reference to design document in readme
