### 2023-10-26 muellerj2/netroles#33 @muellerj2
I'm not a fan of linking to an external website I have no relation with, so let me try first make an attempt to set up a Github action that automatically deploys the Javadoc to muellerj2.github.io/netroles whenever a new release tag is created.

> Although software is documented to a satisfactory level, for a Java package API documentation is scattered between DESIGN.md and USAGE.md and not very structured.

I agree that Javadoc is helpful, which is why it's available on Maven Central. My experience is that Javadoc is very helpful to find out details of the code, e.g., if you are looking for a specific functionality in a class or package, as long as one is already somewhat familiar with the structure of the library. However, I find it hard to understand the overall system implemented in a library just from the Javadoc. These other two files were added for this latter purpose. Unfortunately, there does not really seem to be a standard way to make available this kind of documentation: different Java libraries do different things.
