### 2023-10-26 #5903 @muellerj2
@abhishektiwari

Regarding substantive scholarly effort:

It is the case that this was extracted from another repository. I probably wrote the first pieces of code for role analysis that developed into the core of this library around the end of 2016, when I started working on role equivalences as the topic of my PhD, and have kept developing and extending it since then because I needed it for my academic work, some of which has been published (in 2019 and 2022) and some of which still needs to be. The design of the user API went through some revisions, but had mostly taken its current form in the style of functional programming by early 2020.

Originally, I wrote the role equivalence functionality with the intention to make it a module of a larger general-purpose network analysis library, which was supposed to be made open source in the future. However, all other modules are not in a publishable state and their development has been stalled for years. By now, I have to objectively say the original project is essentially abandoned (although I'm not ready yet to give up on it completely). So I extracted the role analysis module from this larger repository to publish it on GitHub, omitting everything which is unusable or not practically relevant to perform role equivalence analysis, and the outcome is this library.
