### 2023-10-26 #5903 @muellerj2
@leifeld

> The paper itself feels like it lacks more examples and a conclusion, but I guess that's the desired format at JOSS, and it would only duplicate the `USAGE.md` file anyway and potentially go beyond the word limit.

My impression is that papers on R packages in JOSS often contain usage examples. But before I made the paper submission, I checked the published papers on Java libraries and applications in JOSS, and I only found a single paper containing a code example among all of them, and this code example wasn't any longer than the small example in my paper. I guess the issue is that Java needs more boilerplate and is more wordy than a dynamic language like R, so it makes the paper even longer and almost impossible to include a full code example without exceeding the word limit.

The paper has already reached the word limit, which means any addition of usage examples would substantially exceed it. So I guess it will have to be @fboehm's call whether some usage examples could be added. If so, I would try to work in some shorter versions of the usage examples in the `USAGE.md` file.

>  I know @muellerj2 is the main author of visone at this point but just wanted to double-check in line with the review check list.

I wouldn't say I'm the main author of `visone` -- there are people who have contributed much more to `visone2` than me. I prefer to say that I'm the maintainer of the code base, and I took over this responsibility in 2019.

> One question I have is whether @muellerj2 wrote all the code or if the co-author on the Social Networks paper or any of the other visone authors contributed anything.

The co-author of the Social Networks didn't contribute any code and wasn't involved in the software design of the role analysis code (beyond the ideas and concepts we published on in Social Networks). It also includes none of the code of the `visone1` or `visone2` projects, so there is no code in there that was written by anyone who published a paper on prior versions of `visone`. Obviously, though, some design choices are based on rough ideas and concepts that have been discussed in the research group over many years.

Originally, the role analysis core of this library was supposed to become a module of the third iteration of `visone`, which aimed to redevelop `visone` from scratch and was intended to be released as open source software. (Due to its heavy reliance on a commercial graph library, it is not possible to make the predecessor project `visone2` open source or even use its code as a reasonable basis for something that is supposed to become an open source project, besides the large amounts of questionable legacy code it accumulated over two decades.) However, this project has been stalled for years, is essentially abandoned, and includes no mature component now other than the role analysis module that I developed on my own for my PhD work.

I extracted the role analysis module plus a minimal set of utility classes from the `visone3` repository needed to make it run and a bit easier to use, and this became the `netroles` library. I omitted any other functionality that was implemented during the `visone3` project, which means the `netroles` library is only geared towards role equivalence analysis and not a general purpose network analysis library as was the aim of the `visone3` project. I kept the directory structure the same, though, to make it easier for me to apply my changes to both repositories and perhaps simplify the transition in the unlikely case `visone3` itself will be published at some point.

The core of this library -- the role analysis module -- was written by me alone and no one else. But while I have made additions and substantive changes to the utility classes and wrote large portions of the test code and almost all of the documentation for them, many of the utility classes used by the library (specifically containers, network representation and IO) were originally designed and written by someone else. I contacted this person before I published the library on GitHub and asked him as well whether he wanted to appear as a contributor to the library, but he declined. However, if he tells me in the future that he would like his name to appear, I will add it to the repository -- after all, it is his right to decide if and how he is named as a contributor.

Still, I wrote most of the code in the repository (at least 80% of the lines of code, likely quite a bit more but I didn't perform an exact count) and all the code for the library's core role analysis functionality. Only the utility classes contain code not written by me, but these classes could be replaced by similar functionality available in other common Java packages (although this would take quite a bit of rewriting effort since the role equivalence analysis code makes widespread use of them).
