### 2023-10-26 #5903 @leifeld
@fboehm I'm happy with these explanations and recommend publishing the paper.

Thanks for these additional explanations, @muellerj2. As an aside, please don't let visone die. The majority of the [~ 300 publications](https://github.com/leifeld/dna/blob/master/build/bibliography.md) that use [Discourse Network Analyzer](https://github.com/leifeld/dna) benefitted from it.
