### 2023-10-26 #5903 @abhishektiwari
+1 to other reviewers' feedback.

@fboehm completed review from my end. Only open ticket is [Javadoc](https://github.com/muellerj2/netroles/issues/33), but that is not a blocker. I look forward to seeing more research using netroles.
