### 2023-10-29 2b1c8ff
move functional interfaces to dedicated package in lang-api (#39)
