### 2023-10-29 107bddd
Add workflow building release binaries from a git tag
