### 2023-10-30 043d554
extend Javadoc package and class descriptions
adds some example code and more explanations on how API packages or
classes are supposed to be used
