+++ modules/roles-api/src/main/java/ch/ethz/sn/visone3/roles/blocks/package-info.java
@@ -21,24 +21,86 @@
  * <p>
  * This package is organized as follows:
  * <ul>
- * <li>The Operators class offers methods for composing operators in a
- * data-flow-like manner. The operators can be composed sequentially (series(),
- * compose*()) or in parallel working on the same initial input followed by a
- * reduction operation (parallel()).</li>
- * <li>The RoleOperators class provides builders to construct operators for
- * established notions of role on the lattices of equivalences, rankings and
- * binary relations. Moreover, it also offers basic operators working on these
- * lattices.</li>
- * <li>The DistanceOperators class provides builders to construct operators
- * producing distances based on established notions of role, as well as some
- * basic operators working on distances.</li>
- * <li>The Reducers class provides reduction operations for equivalences,
- * rankings, binary relations and distances. These are intended to be used in
- * conjunction with the Operators.parallel() methods.</li>
- * <li>The Converters class provides operators representing conversion
- * operations between the different kinds of role structure representations
- * (equivalence, ranking, binary relation) and distance matrices.
+ * <li>The {@link ch.ethz.sn.visone3.roles.blocks.Operators} class offers
+ * methods for composing operators in a data-flow-like manner. The operators can
+ * be composed sequentially (series(), compose*()) or in parallel working on the
+ * same initial input followed by a reduction operation (parallel()).</li>
+ * <li>The {@link ch.ethz.sn.visone3.roles.blocks.RoleOperators} class provides
+ * builders to construct operators for established notions of role on the
+ * lattices of equivalences, rankings and binary relations. Moreover, it also
+ * offers basic operators working on these lattices.</li>
+ * <li>The {@link ch.ethz.sn.visone3.roles.blocks.DistanceOperators} class
+ * provides builders to construct operators producing distances based on
+ * established notions of role, as well as some basic operators working on
+ * distances.</li>
+ * <li>The {@link ch.ethz.sn.visone3.roles.blocks.Reducers} class provides
+ * reduction operations for equivalences, rankings, binary relations and
+ * distances. These are mainly intended to be used in conjunction with the
+ * Operators.parallel() methods to combine several structures of the same type
+ * into a single result.</li>
+ * <li>The {@link ch.ethz.sn.visone3.roles.blocks.Converters} class provides
+ * operators representing conversion operations between the different kinds of
+ * role structure representations (equivalence, ranking, binary relation) and
+ * distance matrices.
  * <li>
  * </ul>
+ * 
+ * <p>
+ * For example, a role operator that applies the notion of regular equivalence
+ * in both edge directions on a network can be defined as follows using the APIs
+ * of {@link ch.ethz.sn.visone3.roles.blocks.Operators}
+ * {@link ch.ethz.sn.visone3.roles.blocks.RoleOperators} and
+ * {@link ch.ethz.sn.visone3.roles.blocks.Reducers}:
+ * 
+ * <pre>
+ * Network network = ...;
+ * NetworkView<?, ?> outgoingView = NetworkView.fromNetworkRelation(network, Direction.OUTGOING);
+ * NetworkView<?, ?> incomingView = NetworkView.fromNetworkRelation(network, Direction.INCOMING);
+ * // outgoing direction only 
+ * RoleOperator&lt;ConstMapping.OfInt&gt; outgoingRegularOp = RoleOperators.EQUIVALENCE.regular()
+ *   .of(outgoingView).make();
+ * // incoming direction only 
+ * RoleOperator&lt;ConstMapping.OfInt&gt; incomingRegularOp = RoleOperators.EQUIVALENCE.regular()
[…]
