### 2023-10-30 3ba8538
extend and fix javadoc
