+++ modules/roles-impl/src/main/java/ch/ethz/sn/visone3/roles/impl/blocks/factories/package-info.java
@@ -15,6 +15,11 @@
  * along with netroles.  If not, see <http://www.gnu.org/licenses/>.
  */
 /**
- * Implementations of factories for role operators.
+ * Implementations of factories for role operators. This package provides the
+ * actual implementation of the role operator definition pipeline offered in the
+ * {@link ch.ethz.sn.visone3.roles.blocks.RoleOperators} class. The algorithms
+ * in the {@link ch.ethz.sn.visone3.roles.impl.algorithms} class are also
+ * wrapped by suitable role operator classes (which in turn are mostly specified
+ * as anonymous classes in the corresponding factory implementation classes).
  */
 package ch.ethz.sn.visone3.roles.impl.blocks.factories;
