+++ modules/lang-impl/src/main/java/ch/ethz/sn/visone3/lang/impl/mappings/package-info.java
@@ -15,6 +15,9 @@
  * along with netroles.  If not, see <http://www.gnu.org/licenses/>.
  */
 /**
- * Implementation of array-like mappings and lists mainly for primivie types.
+ * Implementation of array-like mappings and lists mainly for primitive types.
+ * The mappings and lists can be created using methods in
+ * {@link ch.ethz.sn.visone3.lang.Mappings}, and corresponding collectors are
+ * made available in {@link ch.ethz.sn.visone3.lang.PrimitiveCollectors}.
  */
 package ch.ethz.sn.visone3.lang.impl.mappings;
