+++ modules/roles-impl/src/main/java/ch/ethz/sn/visone3/roles/impl/blocks/reducers/eq/package-info.java
@@ -15,6 +15,7 @@
  * along with netroles.  If not, see <http://www.gnu.org/licenses/>.
  */
 /**
- * Reduction operations for equivalences.
+ * Reduction operations for equivalences. The operators are made available in
+ * the {@link ch.ethz.sn.visone3.roles.blocks.Reducers} class.
  */
 package ch.ethz.sn.visone3.roles.impl.blocks.reducers.eq;
