### 2023-10-30 b1773e7
extend Javadoc package and class descriptions (#41)
Adds some example code and more explanations on how API packages or classes are supposed to be used. Also briefly describes what parts of the API various implementation packages implement.
