### 2023-10-30 f922003
reduce standard amount of log output when running tests (#42)
