### 2023-10-30 99c4cba
further reduce log test output (#43)
