+++ README.md
@@ -61,12 +61,12 @@ dependencies {
 ```
 
 Alternatively, you can add the following artifacts (available on the Releases page) to your Java project manually:
-* netroles-api-*.*.jar
-* netroles-engine-*.*.jar (only needed at runtime)
-* netroles-networks-api-*.*.jar
-* netroles-networks-engine-*.*.jar (only needed at runtime)
-* netroles-io-api-*.*.jar (optional IO support)
-* netroles-io-engine-*.*.jar (optional IO support, only needed at runtime)
+* netroles-api-\*.\*.jar
+* netroles-engine-\*.\*.jar (only needed at runtime)
+* netroles-networks-api-\*.\*.jar
+* netroles-networks-engine-\*.\*.jar (only needed at runtime)
+* netroles-io-api-\*.\*.jar (optional IO support)
+* netroles-io-engine-\*.\*.jar (optional IO support, only needed at runtime)
 
 In addition, netroles uses [SLF4J](https://www.slf4j.org/) for logging purposes, so you have to add SLF4J and an SLF4J binding to some logging library to your build path. If you also add the optional IO support to your project, you should further add the following external libraries:
 * [Univocity parsers](https://github.com/uniVocity/univocity-parsers)
@@ -78,10 +78,10 @@ In addition, netroles uses [SLF4J](https://www.slf4j.org/) for logging purposes,
 > 
 > You can make use of another library's network data structure than the default one bundled with netroles if you prefer to do so. To achieve this, you have to write an adapter implementing the NetworkView interface in package ch.ethz.sn.visone3.roles.position. You can then use this adapter to specify role concepts on these networks.
 
-## Working on the code
 
-If you would like to work directly on the library's source code, fork this repository and download the library's source. Finally, you can import it into your favorite IDE for Java programming. The project uses Gradle as its build tool, so you might have to install a Gradle plugin to import the project easily into your IDE.
+## Tutorial
 
+You can find a tutorial on the library presenting some usage examples [here](./docs/USAGE.md).
 
 ## Scientific background and design
 
@@ -90,9 +90,27 @@ The design document [DESIGN.md](./docs/DESIGN.md) explains:
 * The theoretical background of this library, including the relevant ideas from the role analysis literature, in the area of Social Network Analysis.
 * The library's design principles.
 
-## Example usage
 
-You can find a tutorial on the library presenting some usage examples [here](./docs/USAGE.md).
+## API Documentation
+
+For each release, corresponding Javadoc documentation is made available on Maven Central and can be accessed on the external website [javadoc.io](javadoc.io/doc/io.github.muellerj2).
+
+You can follow the links below to view the Javadoc of each API module on javadoc.io:
+* [netroles-lang-api](https://javadoc.io/doc/io.github.muellerj2/netroles-lang-api/)
+* [netroles-networks-api](https://javadoc.io/doc/io.github.muellerj2/netroles-networks-api/)
+* [netroles-api](https://javadoc.io/doc/io.github.muellerj2/netroles-api/)
+* [netroles-io-api](https://javadoc.io/doc/io.github.muellerj2/netroles-io-api/)
+
+Besides the API modules, netroles also includes engine modules that implement the library's API. Users of netroles usually do not need to interact with these modules directly. But if you wish to access some feature of an engine module directly, you can browse the Javadoc for these modules as well:
+* [netroles-lang-engine](https://javadoc.io/doc/io.github.muellerj2/netroles-lang-engine/)
+* [netroles-networks-engine](https://javadoc.io/doc/io.github.muellerj2/netroles-networks-engine/)
+* [netroles-engine](https://javadoc.io/doc/io.github.muellerj2/netroles-engine/)
[…]
