### 2023-10-31 941879e
Extend lang-api package description and remove package overview from design.md
