+++ docs/DESIGN.md
@@ -140,105 +140,3 @@ This dataflow-based design presents a number of benefits, particularly in achiev
 * **Optimization opportunities**: The dataflow-based design lends itself to a few natural optimizations. Currently, the library exploits the following optimizations to some degree:
     * ***Constant folding and caching***: If possible, operations are ordered such that operators with constant output are folded and the result of the folding is cached.
     * ***Lazy evaluation***: In some cases, the computation of intermediate results is postponed or only performed to the degree subsequent operators need these results.
-
-## Package overview
-
-Below you find a general overview of the API packages in the netroles library. In the outline, we cover the packages roughly in the order in which general users are likely to interact with them. For a more detailed description of the package contents, please consult the Javadoc documentation available in Maven Central.
-
-### Commonly used packages
-
-#### ch.ethz.sn.visone3.lang (lang-api module)
-
-This package implements general containers and algorithms, mainly for use with primitive types like integers.
-
-* Mappings are the most common data structure in the library. They are fixed-size lists and store values of primitive types compactly (i.e., unboxed). There are read-only (`ConstMapping`) and writable (`Mapping`) mappings, as well as those with variable size (`PrimitiveList`) that function similar to `java.util.List`. New Mappings can be created using methods of the `Mappings` class.
-* Algorithms for permuting, mapping or counting sort are availab in the `PrimitiveCollections` class.
-* The `PrimitiveCollector` interface describes primitive versions of the `Collector` interface in the Java stream API. The `PrimitiveCollectors` class provides collectors for use with the primitive streams `IntStream`, `LongStream` and `DoubleStream`.
-* Some other data structures for primitive data types can be obtained from the `PrimitiveContainers` class.
-
-#### ch.ethz.sn.visone3.networks (networks-api module)
-
-This package provides a default representation for networks.
-* The `Network` interface gives access to three different perspectives: A directed graph (interface `DirectedGraph`), an undirected graph (`UndirectedGraph`) an a relational (`Relation`) perspective. Generally, all networks support the relational perspective, but only one of the graph perspectives in line with the network's directedness.
-* New networks can be created by using builders (of interface `NetworkBuilder`) obtained from the `NetworkProvider` class, or by supplying the corresponding adjacency matrices to methods provided by the `MatrixSource` class.
-
-#### ch.ethz.sn.visone3.roles.blocks (roles-api module)
-
-This is the main package for role analysis. It provides operators to establish roles according to commonly used notions of role equivalence, as well as utilities to produce more complex operators through composition.
-
-* The `Operators` class offers the following generic methods for composing operators:
-    * The `compose*()` family of methods yield operators applying one operator to the output of another.
-    * The` series()` methods yield operators that apply several operators in series, feeding the output of one operator as input to the next one (cf. series circuits).
-   * The `parallel()` methods yield operators that apply several operators to the same input and then combine their results using a reduction operation (cf. parallel circuits). Some common reduction operators are provided in the `Reducers` class.
-* The `RoleOperators` class provides factories to construct (mostly unidirectional) operators for commonly used role notions. All of these are offered in three versions depending on the kind of representation of role structure: equivalence, ranking or binary relation. For example, the factory for regular roles on the lattice of equivalences can be accessed through `RoleOperators.EQUIVALENCE.regular()`.
-    * The `EQUIVALENCE.basic()`, `RANKING.basic()` and `BINARYRELATION.basic()` methods additionally provide some generic basic operators acting on equivalences, rankings and binary relations.
-    * The `regular()` family of factories lead to operators for ecological, perfect and regular roles.
[…]
