### 2023-11-01 b956bff
Extend lang-api, io-api package description and remove package overview (#46)
from design.md

* add more descriptive overview of package contents to packages
* state in API documentation what formats the default implementation supports
* remove the package overview (as its contents have been integrated into javadoc)
