### 2023-11-01 4c9b104
fix missed change in design document (#48)
