### 2023-11-01 33486dc
output human-readable heading for implNote javadoc tag (#49)
drive-by change: fix strange capitalization in IOProvider implNote
