### 2023-11-02 muellerj2/netroles#33 @muellerj2
I made a few improvements to the Javadoc documentation. I extended the Javadoc descriptions of some classes and packages, adding code examples or giving some higher-level explanations of the package contents.

These changes made the package overview section in the DESIGN.md basically superfluous. Besides, it's likely that I would forget to update this section to reflect future changes. That's why I have removed this section from the document.

Another minor change is that the generated Javadoc in the next release will add links to https://docs.oracle.com/javase/8/docs/api/ for JDK classes.

Regarding the main request of this issue: In the end, I decided to not host the Javadoc on GitHub pages (for now) but to link to javadoc.io as suggested. PR #45 adds a section in README.md that links to javadoc.io for all library modules. I did not add the badges, because I think eight badges saying "javadoc" next to each other is more confusing than helpful. When I added this section, I also reordered and reworded the sections referencing USAGE.md and DESIGN.md in the readme.

You can view the branch of PR #45 with the updated readme here: https://github.com/muellerj2/netroles/tree/link-javadoc-in-readme . Please tell me if you have further suggestions for improvement. If not, I will merge the PR as-is, closing this issue.
