+++ paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 
 # Summary
 
-An important problem in the analysis of networks is structural similarity. It has commonly been expressed in terms of role equivalences, which have often been considered formalizations of the concepts of social role and position as discussed by @l-sm-36, @m-stss-57 and @n-tss-57 in their analyses of social structure [@be-npsna-92]. Role equivalences are based on the idea that actors are equivalent or play the same role if they form ties to similar others in similar ways. For example, the role of "doctor" is characterized by a set of ties to others playing related roles like "nurses", "patients" and "colleagues". However, this equivalence idea has been interpreted by different authors in different ways, resulting in the proposition of diverse definitions of role equivalence such as structural [@lw-seisn-71] or regular equivalence [@wr-gshnr-83]. The `netroles` library provides implementations of many established notions of role equivalence, but more importantly offers a unified approach to role equivalence analysis that generalizes beyond the classic role equivalences, allowing users to express more complex kinds of role notions suitable for networks with multiple relations and attributes.
+An important problem in the analysis of networks is structural similarity. It has commonly been expressed in terms of role equivalences, which have often been considered to formalize the concepts of social role and position as discussed by @l-sm-36, @m-stss-57 and @n-tss-57 in their analyses of social structure [@be-npsna-92]. Role equivalences are based on the idea that actors are equivalent or play the same role if they form ties to similar others in similar ways. For example, the role of "doctor" is characterized by a set of ties to others playing related roles like "nurses", "patients" and "colleagues". However, this equivalence idea has been interpreted by different authors in different ways, resulting in the proposition of diverse definitions of role equivalence such as structural [@lw-seisn-71] or regular equivalence [@wr-gshnr-83]. The `netroles` library provides implementations of many established notions of role equivalence, but more importantly offers a unified approach to role equivalence analysis that generalizes beyond the classic role equivalences, allowing users to express more complex kinds of role notions suitable for networks with multiple relations and attributes.
 
 # Statement of need
 
