### 2023-11-02 #5903 @muellerj2
I will build and publish the new release after merging the last open PR from the review into the main branch.

@leifeld I would like to make some promises on visone, but honestly, I can't even promise that I will still be in academia a year from now. That said, we haven't abandoned visone yet and are thinking about ways to revitalize the project.
