### 2023-11-03 0b40020
update readme: link to Javadoc API docs and reword sections for tutorial and design document (#45)
