+++ README.md
@@ -19,13 +19,13 @@ Netroles is available in Maven Central. You can add it by using your preferred d
     <groupId>io.github.muellerj2</groupId>
     <artifactId>netroles-api</artifactId>
     <version>x.y</version>
-	<scope>compile</scope>
+    <scope>compile</scope>
 </dependency>
 <dependency>
     <groupId>io.github.muellerj2</groupId>
     <artifactId>netroles-engine</artifactId>
     <version>x.y</version>
-	<scope>runtime</scope>
+    <scope>runtime</scope>
 </dependency>
 ```
 
@@ -37,13 +37,13 @@ Netroles includes its own implementation of network structures. Optionally, you 
     <groupId>io.github.muellerj2</groupId>
     <artifactId>netroles-io-api</artifactId>
     <version>x.y</version>
-	<scope>compile</scope>
+    <scope>compile</scope>
 </dependency>
 <dependency>
     <groupId>io.github.muellerj2</groupId>
     <artifactId>netroles-io-engine</artifactId>
     <version>x.y</version>
-	<scope>runtime</scope>
+    <scope>runtime</scope>
 </dependency>
 ```
 
@@ -61,12 +61,12 @@ dependencies {
 ```
 
 Alternatively, you can add the following artifacts (available on the Releases page) to your Java project manually:
-* netroles-api-*.*.jar
-* netroles-engine-*.*.jar (only needed at runtime)
-* netroles-networks-api-*.*.jar
-* netroles-networks-engine-*.*.jar (only needed at runtime)
-* netroles-io-api-*.*.jar (optional IO support)
-* netroles-io-engine-*.*.jar (optional IO support, only needed at runtime)
+* netroles-api-\*.\*.jar
+* netroles-engine-\*.\*.jar (only needed at runtime)
+* netroles-networks-api-\*.\*.jar
+* netroles-networks-engine-\*.\*.jar (only needed at runtime)
+* netroles-io-api-\*.\*.jar (optional IO support)
+* netroles-io-engine-\*.\*.jar (optional IO support, only needed at runtime)
 
 In addition, netroles uses [SLF4J](https://www.slf4j.org/) for logging purposes, so you have to add SLF4J and an SLF4J binding to some logging library to your build path. If you also add the optional IO support to your project, you should further add the following external libraries:
 * [Univocity parsers](https://github.com/uniVocity/univocity-parsers)
@@ -78,21 +78,39 @@ In addition, netroles uses [SLF4J](https://www.slf4j.org/) for logging purposes,
 > 
 > You can make use of another library's network data structure than the default one bundled with netroles if you prefer to do so. To achieve this, you have to write an adapter implementing the NetworkView interface in package ch.ethz.sn.visone3.roles.position. You can then use this adapter to specify role concepts on these networks.
 
+
+## Tutorial
+
+The [tutorial](./docs/USAGE.md) presents common usage examples step by step.
+
+## Scientific background and design
+
+The [design document](./docs/DESIGN.md) explains:
+
+* the theoretical background of this library in the area of Social Network Analysis, including the relevant ideas from the role analysis literature.
+* the library's design principles.
+
+
+## API Documentation
+
+For each release, corresponding Javadoc documentation is made available on Maven Central and can be accessed on the external website [javadoc.io](https://javadoc.io/doc/io.github.muellerj2).
+
+You can follow the links below to view the Javadoc of each API module on javadoc.io:
+* [netroles-lang-api](https://javadoc.io/doc/io.github.muellerj2/netroles-lang-api/)
+* [netroles-networks-api](https://javadoc.io/doc/io.github.muellerj2/netroles-networks-api/)
[…]
