### 2023-11-03 9e05b3a
minor change: print explicitly what is the domain size
update USAGE.md accordingly
link to underlying paper at all reference points
