### 2023-11-03 17f6981
minor change: print explicitly what is the domain size (#52)
update USAGE.md accordingly
drive-by change: link to paper at all points it is referenced
