### 2023-11-03 34e1173
prepare release of version 1.0 (#53)
set version number
align Maven descriptions of API modules
clean before performing the build using Gradle (to make unintended side effects of Gradle caches less likely)
