### 2023-11-04 475cb75
updated bibliography references
