+++ paper.bib
@@ -18,13 +18,16 @@
   langid = {english},
 }
 
-@incollection{bbbcgklw-vsvsna-02,
-  title = {Visone Software for Visual Social Network Analysis},
+@inproceedings{bbbcgklw-vsvsna-02,
+  title = {{{visone}} Software for Visual Social Network Analysis},
   booktitle = {Graph {{Drawing}}},
+  booktitleaddon = {{9th International Symposium, GD 2001}},
   author = {Baur, Michael and Benkert, Marc and Brandes, Ulrik and Cornelsen, Sabine and Gaertler, Marco and K{\"o}pf, Boris and Lerner, J{\"u}rgen and Wagner, Dorothea},
-  editor = {Goos, Gerhard and Hartmanis, Juris and {van Leeuwen}, Jan and Mutzel, Petra and J{\"u}nger, Michael and Leipert, Sebastian},
+  editor = {Mutzel, Petra and J{\"u}nger, Michael and Leipert, Sebastian},
   year = {2002},
-  volume = {2265},
+  venue = {Vienna, Austria},
+  series = {Lecture {{Notes}} in {{Computer Science}}},
+  number = {2265},
   pages = {463--464},
   publisher = {{Springer}},
   address = {{Berlin, Heidelberg}},
@@ -99,7 +102,7 @@
   title = {Pajek \textemdash{} Analysis and Visualization of Large Networks},
   booktitle = {Graph {{Drawing Software}}},
   author = {Batagelj, Vladimir and Mrvar, Andrej},
-  editor = {Farin, Gerald and Hege, Hans-Christian and Hoffman, David and Johnson, Christopher R. and Polthier, Konrad and J{\"u}nger, Michael and Mutzel, Petra},
+  editor = {J{\"u}nger, Michael and Mutzel, Petra},
   year = {2004},
   pages = {77--103},
   publisher = {{Springer}},
