### 2023-11-04 #5903 @muellerj2
Version of software release: 1.0

Zenodo DOI link: https://doi.org/10.5281/zenodo.10070537

I checked/set author and affiliation names in the paper and on Zenodo, and made some minor corrections to two references.

If there is anything else I need to do, please tell me.
