+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-junit = "5.10.0"
+junit = "5.10.1"
 slf4j = "2.0.9"
 log4j2 = "2.21.1"
 
