### 2023-12-11 #5903 @fboehm
@muellerj2 - here are a few questions for paper.pdf:

1. On line 84, should `visone` be all lower case letters, like it now is? Or do some of the letters need capitalization?
2. On line 102, is "sna" correctly uncapitalized?

I'm guessing that you've capitalized (or uncapitalized) correctly, but i just want to make sure.
