+++ paper.bib
@@ -110,6 +110,18 @@
   doi = {10.1007/978-3-642-18638-7_4},
 }
 
+@incollection{bw-vavsn-04,
+  title = {{{visone}} \textemdash{} Analysis and Visualization of Social Networks},
+  booktitle = {Graph {{Drawing Software}}},
+  author = {Brandes, Ulrik and Wagner, Dorothea},
+  editor = {J{\"u}nger, Michael and Mutzel, Petra},
+  year = {2004},
+  pages = {321--340},
+  publisher = {{Springer}},
+  address = {{Berlin, Heidelberg}},
+  doi = {10.1007/978-3-642-18638-7_15}
+}
+
 @misc{cdb-s-23,
   title = {{sbm}: {{Stochastic}} Blockmodels},
   author = {Chiquet, Julien and Donnet, Sophie and Barbillon, Pierre},
