### 2023-12-11 #5903 @muellerj2
@fboehm  Thank you for handling the paper! Regarding the two capitalization issues you raised, I made one minor change, but see my comments below.

> On line 84, should visone be all lower case letters, like it now is? Or do some of the letters need capitalization?

The name of the software `visone` is typically not capitalized. That said, I checked the original publication and it is capitalized in the proceedings' table of contents for some reason. But I strongly suspect this capitalization was not decided by the authors, since the rest of the article title is also very strangely formatted in the table of contents. On the other hand, the actual article uses the logo of the software (which is all lower case) in the title. So there are arguments either way, but I personally think lower case was originally intended.

After pondering for a while, I decided to do the following:

* I added a citation to another article on visone. That's probably the better article on visone anyway and the PDF version of it is very clear in that the software name is supposed to be written in lower case.
* But I kept the other citation as well, mainly because it gives credit to many of the original developers of visone, while the newly referenced article was only authored by the leadership of the software project. I didn't change the capitalization in this reference, so visone is still written in lower case.

This slightly affects the formatting in lines 22-26 and adds lines 96-98, consequently shifting the following bibliographical references a bit further down. I did not touch anything else, so the rest is unchanged. I hope doing this change is fine.

> On line 102, is "sna" correctly uncapitalized?

Yes. "SNA" (capitalized) is the acronym for the field Social Network Analysis, while `sna` (uncapitalized) is the R package providing tools for Social Network Analysis. Furthermore, sna is also written in lower case in the PDF version of the referenced article.
