+++ .github/workflows/gradle_build.yml
@@ -39,7 +39,7 @@ jobs:
       with:
         arguments: build
     - name: Upload reports
-      uses: actions/upload-artifact@v3
+      uses: actions/upload-artifact@v4
       with:
         name: build-reports
         path: build/reports/*
