### 2024-01-02 0010b8a
Bump org.slf4j:slf4j-api from 2.0.9 to 2.0.10
Bumps org.slf4j:slf4j-api from 2.0.9 to 2.0.10.

---
updated-dependencies:
- dependency-name: org.slf4j:slf4j-api
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
