+++ gradle/libs.versions.toml
@@ -1,6 +1,6 @@
 [versions]
 junit = "5.10.1"
-slf4j = "2.0.9"
+slf4j = "2.0.10"
 log4j2 = "2.22.0"
 
 [libraries]
