+++ gradle/libs.versions.toml
@@ -1,7 +1,7 @@
 [versions]
 junit = "5.10.1"
 slf4j = "2.0.10"
-log4j2 = "2.22.0"
+log4j2 = "2.22.1"
 
 [libraries]
 junit-launcher = { group = "org.junit.platform", name = "junit-platform-launcher" }
