+++ gradle/libs.versions.toml
@@ -1,6 +1,6 @@
 [versions]
 junit = "5.10.1"
-slf4j = "2.0.10"
+slf4j = "2.0.11"
 log4j2 = "2.22.1"
 
 [libraries]
