### 2024-01-15 #5903 @muellerj2
@fboehm No problem.

The paper was only changed by the following commit compared to the previous version you already checked: https://github.com/muellerj2/netroles/commit/3a3247747312560ede2770deb3562a8cfab31518

This should make it easier for you to decide where to take a closer look in the updated paper.
