### 2024-01-17 #5903 @danielskatz
@muellerj2 - I've suggested some minor changes in https://github.com/muellerj2/netroles/pull/64 - please merge this, or let me know what you disagree with, then we can proceed to acceptance and publication
