### 2024-01-17 #5903 @muellerj2
@danielskatz I'm fine with these changes and applied them to the paper.
