### 2024-01-17 #5903 @danielskatz
Congratulations to @muellerj2 (Julian Müller) on your publication!!

JOSS depends on volunteers and couldn't do this without you
